:truck: move consts.ts
