+++ src/pages/Editor/Description/AnnualPlanEditor.tsx
@@ -3,7 +3,7 @@ import { useState } from "react"
 import { PortalButton } from "../../../components/common/Button"
 import { EditorBase } from "../../../components/common/Editor/EditorBase"
 import { TitleArea } from "../../../components/global/Header/TitleArea"
-import { PADDING_BEFORE_FOOTER } from "../../../static/consts"
+import { PADDING_BEFORE_FOOTER } from "../../../utils/consts"
 import type { StateDispatch } from "../../../types/utils"
 
 const months = [4, 5, 6, 7, 8, 9, 10, 11, 12, 1, 2, 3]
