+++ src/pages/Editor/Description/DetailInformation/DetailInformationEditor.tsx
@@ -3,7 +3,7 @@ import { useState } from "react"
 import { PortalButton } from "../../../../components/common/Button"
 import { EditorBase } from "../../../../components/common/Editor/EditorBase"
 import { TitleArea } from "../../../../components/global/Header/TitleArea"
-import { PADDING_BEFORE_FOOTER } from "../../../../static/consts"
+import { PADDING_BEFORE_FOOTER } from "../../../../utils/consts"
 import type { StateDispatch } from "../../../../types/utils"
 import { ActivityEditor } from "./ActivityEditor"
 import { DatetimeEditor, DatetimeItem } from "./DatetimeEditor"
