+++ src/pages/Editor/Description/IntroductionVideoEditor.tsx
@@ -22,7 +22,7 @@ import { BsQuestionCircle } from "react-icons/bs"
 import { PortalButton } from "../../../components/common/Button"
 import { EditorBase } from "../../../components/common/Editor/EditorBase"
 import { TitleArea } from "../../../components/global/Header/TitleArea"
-import { PADDING_BEFORE_FOOTER } from "../../../static/consts"
+import { PADDING_BEFORE_FOOTER } from "../../../utils/consts"
 import type { StateDispatch } from "../../../types/utils"
 
 function parseVideoId(
