+++ src/pages/Editor/Editors.tsx
@@ -1,7 +1,7 @@
 import { VStack } from "@chakra-ui/react"
 import { EditorMenu } from "../../components/common/Editor/EditorMenu"
 import { TitleArea } from "../../components/global/Header/TitleArea"
-import { PADDING_BEFORE_FOOTER } from "../../static/consts"
+import { PADDING_BEFORE_FOOTER } from "../../utils/consts"
 
 export const Editors: React.VFC<{}> = () => {
   return (
